\name{GONNET}
\alias{GONNET}
\alias{GONNETBF}

\docType{data}
\title{
  The GONNET AA matrix
}
\description{
  The GONNET mutation matrix and background frequency.
}
\usage{
data(GONNET)
data(GONNETBF)
}
\format{
  A 20*20 numeric matrix.
  
  A 20 numeric vector.
}

\source{
  Gonnet, G.H., Cohen, M.A., and Benner, S.A. (1992). Exhaustive matching of the entire protein sequence database. Science 256, 1443-1445.
}

\examples{
data(GONNET)
data(GONNETBF)
}

\keyword{datasets}
